+++ R/pin_remove.R
@@ -1,4 +1,3 @@
-
 #' Remove Pin
 #'
 #' Unpins the given named pin from the given board.
@@ -6,22 +5,16 @@
 #' @param name The name for the pin.
 #' @param board The board from where this pin will be removed.
 #'
-#' @details
-#'
-#' Notice that some boards do not support deleting pins, this is the case
-#' for the Kaggle board. For these boards, you would manually have to
-#' remote resources using the tools the board provides.
-#'
 #' @examples
-#' # define temporary board
 #' board <- board_temp()
-#'
-#' # create mtcars pin
 #' pin(mtcars, board = board)
-#'
-#' # remove mtcars pin
 #' pin_remove("mtcars", board = board)
+#'
+#' # ->
+#' board %>% pin_write(mtcars)
+#' board %>% pin_delete("mtcars")
 #' @export
+#' @keywords internal
 pin_remove <- function(name, board) {
   board <- board_get(board)